fix(v1.6.1): post-review critical/major/accidental/optimizations batch
C2: isLowValueCandidate rejects ReclassifiedFrom=blocked (cross-pass double-promotion fix)
M6: Rule 4 temporal bucket blocked -> junk
M7: Catchall rule for unmatched low-value PRs
M1: DynamicTargetConfig.Enabled defaults true for zero-valued structs
M3: Empty-state guard in ReclassificationSection.Render
M4: LoadCollapseImpactSection returns zero-section instead of error
M5+M8: First-match-wins + boundary value tests
A1: Rule 4 ReclassifiedFrom captures actual original category
A2: serve.go copies DynamicTarget config
A3: --target-ratio upper bound validation (>1.0 rejected)
O3: Bootstrap _coerce_dataclass handles missing keys
O2: Remove bestScore no-op assignment
O1: Verified gate17 fresh per-PR
O4: Merge duplicate TS AnalysisResponse declarations
O5: duplicate_synthesis field present in Python models
C1: TODO.md uses ReviewCategory constants with two-level abstraction docs

+++ internal/review/recovery.go
@@ -217,6 +217,7 @@ func isCIFailing(ciStatus string) bool {
 }
 
 // isRecentPush returns true if the PR was pushed within the specified number of days.
+// A PR is NOT recent if it was last updated >= withinDays days ago.
 func isRecentPush(updatedAt string, withinDays int) bool {
 	if updatedAt == "" {
 		return false
@@ -225,7 +226,9 @@ func isRecentPush(updatedAt string, withinDays int) bool {
 	if err != nil {
 		return false
 	}
-	return time.Since(t).Hours()/24 < float64(withinDays)
+	// Use integer division to avoid floating point precision issues at boundaries.
+	// A PR is recent if days since update < withinDays, i.e., NOT recent if >= withinDays.
+	return int(time.Since(t).Hours()/24) < withinDays
 }
 
 // isMergeableUnknown returns true if the mergeable status is unknown or empty.
